Move the Context class to ThreadedHttpRequestHandler
- Add new ctor that takes a Context
